implement rabl on item

+++ db/migrate/20120315180656_add_vendor_id_to_items.rb
@@ -1,5 +1,5 @@
 class AddVendorIdToItems < ActiveRecord::Migration
   def change
-    add_column :items, :vendor_id, :long
+    add_column :items, :vendor_id, :integer
   end
 end
