+++ db/migrate/20120315184032_add_user_to_order.rb
@@ -1,5 +1,5 @@
 class AddUserToOrder < ActiveRecord::Migration
   def change
-    add_column :orders, :user_id, :long
+    add_column :orders, :user_id, :integer
   end
 end
