+++ db/migrate/20120319195430_add_many_to_many_on_order_and_item.rb
@@ -1,6 +1,6 @@
 class AddManyToManyOnOrderAndItem < ActiveRecord::Migration
   def change
-    add_column :order_items, :order_id, :long
-    add_column :order_items, :item_id, :long
+    add_column :order_items, :order_id, :integer
+    add_column :order_items, :item_id, :integer
   end
 end
